+++ MovieQuizTests/ArrayTests.swift
@@ -8,7 +8,7 @@
 import XCTest
 @testable import MovieQuiz
 
-class ArrayTests: XCTestCase {
+final class ArrayTests: XCTestCase {
     func testGetValueInRange() throws {
         // Given
         let array = [1, 1, 2, 3, 5]
